requirements.txt
# Database
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
asyncpg>=0.29

# API
fastapi>=0.115
uvicorn>=0.30

# Validation & Config
pydantic>=2.0
python-dotenv>=1.0
